Use squashfs to contain raw images
We can't simply create a squashfs system image as the file UIDs,
GIDs, and permissions are all wrong -- there is some secret sauce
in make_ext4fs which sets all this up properly. This totally broke
the boot in Jelly Bean as init was refusing to read build.prop due
to its file permissions.

Instead, create a compressed container filesystem for images and put
system.img (along with all the other images) within it.

Change-Id: Ib91de09784413d4e5b053245cca8a0d1333ddc5c
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>