iago: installer backend
This implements the IAGO installer daemon, SYSLINUX, and Droidboot
plug-ins. This has been verified to run end-to-end and produce a
bootable Android installation if all other Iago-enabling patches
are cherry-picked.

The "cli" command line phases are just filling in defaults for now.
Support for a proper default ini overlay and a real command line
interface forthcoming in subsequent patch sets.

Depends-Change-Id: I04782691d4e5ff0db6f68d8d9345a8585cfe5f86
Change-Id: I394203dd5a28af023f5c8fa6d1017602d3361a15
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>