enable command line interface
() iago-default.ini now contains defaults for the non-interactive
installer. Plugins and BoardConfigs can update this.
() Preinit now amends /default.prop with the device used for
installation media. The user won't be able to select this device
to install on.
() Some crude command line interface code added, to provide the
user with a list of options. The default is the first one.
() Partitioner now can ask the user for the destination disk in cli mode
() 'quiet' passed to the kernel so printks don't interfere with
CLI entry
() Plugins don't have to specify ini files if they don't need to
populate any options

Depends-Change-Id: I760878da88ca39bcabc4497383a884be668f57a3
Change-Id: Iecb020dde1d5f627e9a3bd0ce9a074280d7ea508
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>