create a second bootloader partition
During an OTA update, the contents of bootloader are copied to
bootloader2. The update logic then does all its work on that partition.
This ensures that an interruption due to power loss won't brick the
device. Once the operation is complete, the labels and boot flags
of these two partitions are swapped in the GPT.

Change-Id: I366c5adec9ddd300efcf65567ef49d2d0d67664a
Original-Change-Id: I27dc0192b7530233aeb6b38571896646e965dfe5
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>