partitioner: detect existing configurations
The partitioner can now detect the presence of a Windows 8
installation, a previous installation of Android, or a
previously existing EFI system partition. Actions based on these
data to be implemented in forthcoming patches.

The install_id is now a 32-bit value prefixed with "ANDROID!".
This makes it trivial to detect prior installations. The chance
of a collision is still infinitesimal.

Change-Id: I35619f92795d98672fd0b0d4aad9e0ea44518874
Original-Change-Id: I174d045dcfb11f9dea7aa0a9f4c8965869f2d21e
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>