use AOSP boot images for installation media
We'd like to do the same verification and security policies
for the installation media as we do for the installed system.
This will also make it easier to test the secure boot path.
Use the new capability in Gummiboot to load the boot images
from a file.

We create different boot images for the 3 installation options
since (by design) supplemental command line information cannot
be placed in the loader config when loader AOSP boot images.

Change-Id: I1f77eaeefa92a812818d13a9ba35dc7589455ebd
Depends-Change-Id: I97c9bea1da35e548bbb6c44be9e88cb0041c12ba
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>