provisioning image fixes
() Provisioning images always re-create the ESP and partition
table; the code was incorrectly defaulting to the dual boot path.

() newfs_msdos is not a standalone executable and is part of
toolbox. In order not to depend on a full toolbox installation,
drop in the newfs_msdos.c sources from the toolbox repo.
Fortunately this code is self-contained.

() Don't do ntfsresize checks in non-interactive installations;
add whether this is an interactive session to the property store.

() Make it clearer when a fatal assertion has been encountered.

Change-Id: Ibf42a0a01613f6757e8ec073cea39c69229e5633
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>