refactor install-id mechanism
This has been identified as a security risk and should not be
used, as a specially-crafted removable media could be used to
alias the /system partition.

Instead we take a new approach. AOSP already creates by-name
symlinks in the following format:

/dev/block/platform/<controller>/by-name/...

Where the names are pulled directly from the GPT. That leaves
the controller name, which we snarf out of sysfs and set in
the kernel command line. init.rc now has a line:

symlink /dev/block/platform/${ro.boot.disk}/by-name /dev/block/by-name

So we can use by-name links as before.

Issue: ADEV-21
Change-Id: I6e0aaf6026c6670ad4d469ae14a3ebad6a378868
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>