iago/partitioner: Stop partitioner deleting ESP partion
The delete android function didn't get updated to reflect the new
magic ids. This caused it to delete the ESP partition.

This would only occurr if the installer gets run again, it detects
the new android_bootloader partition, it would delete it in delete_android
but marks it as skip in the later in the execute_dual_boot function.

V2: No functional change, correcting commit message.

Change-Id: Ic8c18db60f018b6e61e99786f40c922f796414f6