force Iago to interact on the VGA console
Eventually Iago will be implemented as a GUI in the Android live
environment, but until then, we need interactive messages to be
sent to the VGA console independent on what device thinks the
console device is. All messages are also copied to the logcat and
kernel logs.

This does mean that the interactive installer won't work over
serial console; such devices that don't have a display should use
the automated installer or a provisioning image.

Change-Id: I470649d5a9722ecbb86f7244f68c260d4257e328
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>