Fix for optional keys in hashmapGetPrintf function
The function hashmapGetPrintf will never return NULL values since it
will die on such values.
But, there are some places in code where the returned value is checked
for NULL (which means that that key is optional, not required).
So, instead of defaulting to NULL and checking for NULL (case which will
never happen), default it to empty string and check it for empty string.
Because of this logic, when a new parameter was added
(partition.swap:guid) but not found in the hashmap iago was failing
because it could not find this key. Even though, according to the code
logic, this key is optional.

Change-Id: Ieac5bbe9abe2c9279b341dcb296b69cbb337bfc4
Signed-off-by: Robert Chiras <robert.chiras@intel.com>